remove dev consoles and packed to uint128 stakeRewardSplit and treasurySplit

+++ contracts/MagicDepositorConfig.sol
@@ -7,16 +7,16 @@ contract MagicDepositorConfig is Ownable {
     event UpdatedConfiguration(uint256 stakeRewardSplit, uint256 treasurySplit, address treasury, address staking);
 
     /** Config variables */
-    uint256 internal stakeRewardSplit; // Proportion of harvest that is going to stake rewards
-    uint256 internal treasurySplit; // Proportion of harvest that goes to the treasury
+    uint128 internal stakeRewardSplit; // Proportion of harvest that is going to stake rewards
+    uint128 internal treasurySplit; // Proportion of harvest that goes to the treasury
     address internal treasury; // Address of the treasury
     address internal staking; // Address of the staking contract
 
     /** ACCESS-CONTROLLED FUNCTIONS */
 
     function setConfig(
-        uint256 _stakeRewardSplit,
-        uint256 _treasurySplit,
+        uint128 _stakeRewardSplit,
+        uint128 _treasurySplit,
         address _treasury,
         address _staking
     ) external onlyOwner {
@@ -38,8 +38,8 @@ contract MagicDepositorConfig is Ownable {
         external
         view
         returns (
-            uint256,
-            uint256,
+            uint128,
+            uint128,
             address,
             address
         )
